+++ HISTORY.rst
@@ -9,6 +9,7 @@ Release History
   `atmelavr <http://docs.platformio.ikravets.com/en/latest/platforms/atmelavr.html>`_
   MCUs (`issue #35 <https://github.com/ivankravets/platformio/issues/35>`_)
 * Fixed urllib3's SSL warning under Python <= 2.7.2  (`issue #39 <https://github.com/ivankravets/platformio/issues/39>`_)
+* Fixed bug with Arduino's USB boards (`issue #40 <https://github.com/ivankravets/platformio/issues/40>`_)
 
 0.9.2 (2014-12-10)
 ------------------